Improve interface of both crates
This commit adds a feature to the `psa-crypto` crate to allow it to be
built without a library backing its API - this leaves only the type
system in place, with none of the callable methods.
The bindgen exports of `psa-crypto-sys` are also restricted to those
strictly necessary to the `psa-crypto` crate.

Signed-off-by: Ionut Mihalcea <ionut.mihalcea@arm.com>