build.rs: Build Mbed TLS with ENABLE_PROGRAMS=OFF, ENABLE_TESTING=OFF.
In building psa-crypto-sys we never execute the Mbed TLS tests, so
there is little point in building them. By not building them we save
CPU cycles and disc space and we make these crates a little more
portable as the tests use C library facilities that are not required
by the library.

Signed-off-by: Edmund Grimley Evans <edmund.grimley-evans@arm.com>