Replace the "no-std" feature with a "std" feature.
Negative feature names are rejected by Clippy: they cause confusion
when several crates depend on the same crate but with different sets
of features.

The default set of features is changed but anyone who specifies
"--no-default-features" or "default-features = false" will need to
change what features they specify, so this is an incompatible change.

Signed-off-by: Edmund Grimley Evans <edmund.grimley-evans@arm.com>