Add a prefix to the library names and to every global symbol.
A collision with another crate, even another crate that uses the same
C library, should now be very unlikely, so the "links" line in
Cargo.toml is removed. It should also be possible for a project to use
multiple versions of this crate.

The libraries are now called libpsa_crypto_0_10_0_shim.a and
libpsa_crypto_0_10_0_mbedcrypto.a and there is a new directory for
them (.../out/llib) which does not contain any other libraries. All
the globally defined symbols have the prefix "psa_crypto_0_10_0_".

Signed-off-by: Edmund Grimley Evans <edmund.grimley-evans@arm.com>