build: make psa-crypto-sys no_std, fix zeroize with no_std
As psa-crypto-sys does not rely on any std features, it should be safe
to completely switch its bindgen options to use core, and mark it as
no_std. This allows this crate to be used in embedded applications such
as TrustedFirmware-M Secure Partitions.

zeroize should not need `alloc` default feature, as this crate does not
make use of Vec or Box and does not derive zeroize on them. Disable all
default features to avoid unnecessary dependency on std.

Tested to provide access to hash functions via `operations` feature in
a TF-M Secure Partition sample.

Signed-off-by: Dmitrii Sharshakov <d3dx12.xx@gmail.com>